requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data Validation and Serialization
pydantic[email]==2.5.0

# Password Hashing and Security
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Authentication (if needed in future)
python-multipart==0.0.6

# Development and Testing (optional)
pytest==7.4.3
httpx==0.25.2
